refactor(galgame): split models store context service
