+++ plugin/plugins/galgame_plugin/context_builder/__init__.py
@@ -1,0 +1,76 @@
+from __future__ import annotations
+
+from .builder import (
+    _CONDENSE_BLOCKING_PUNCTUATION_RE,
+    _CONDENSE_SHORT_LINE_MAX_CHARS,
+    _DIALOGUE_PUNCTUATION_RE,
+    _DIALOGUE_WEAK_PUNCTUATION_RE,
+    _DYNAMIC_WINDOW_DEFAULT_MAX_LINES,
+    _DYNAMIC_WINDOW_DEFAULT_MIN_LINES,
+    _DYNAMIC_WINDOW_DEFAULT_TARGET_TOKENS,
+    _IMPORTANCE_EMOTIONAL_PUNCTUATION_RE,
+    _IMPORTANCE_PLOT_WORDS,
+    _IMPORTANCE_TURN_WORDS,
+    _NON_DIALOGUE_CONTEXT_TOKENS,
+    _OCR_OVERLAY_TEXT_GUARD_SUBSTRINGS,
+    _SUMMARY_MAX_CHARS,
+    _append_limited_with_importance,
+    _append_unique_line,
+    _bounded_summary_text,
+    _build_input_degraded_context,
+    _compact_lines_by_importance,
+    _compact_profile_text,
+    _compute_dynamic_line_limit,
+    _condense_dialogue_batch,
+    _condense_run_key,
+    _context_snapshot_summary_seed,
+    _context_window_bounds,
+    _cross_scene_memory_context,
+    _cumulative_scene_summary,
+    _current_line_entry,
+    _dialogue_context_lines,
+    _dialogue_line_dedupe_key,
+    _ensure_target_line_present,
+    _fixed_character_pov_context,
+    _global_scene_context_window,
+    _is_memory_reader_identifier,
+    _is_ocr_reader_identifier,
+    _line_condense_blocked,
+    _line_importance_score,
+    _llm_refined_summary_from_state,
+    _looks_like_game_dialogue_context_line,
+    _looks_like_ocr_overlay_text,
+    _matching_context_snapshot,
+    _merge_condensed_run,
+    _previous_scene_id_from_state,
+    _previous_summary_from_state,
+    _profile_voice_traits,
+    _recency_ordered_context_lines,
+    _resolve_target_line,
+    _scene_context_hint,
+    _scene_history_dialogue_line_count,
+    _scene_lines,
+    _scene_selected_choices,
+    _scene_summary_seed_with_restored_context,
+    _significant_char_count,
+    _snapshot_for_stable_summary_seed,
+    _strip_importance_score,
+    _summary_mode,
+    build_explain_context,
+    build_fallback_summary,
+    build_local_scene_summary,
+    build_ocr_context_diagnostic,
+    build_suggest_context,
+    build_summarize_context,
+    resolve_effective_current_line,
+)
+
+__all__ = [
+    "build_explain_context",
+    "build_fallback_summary",
+    "build_local_scene_summary",
+    "build_ocr_context_diagnostic",
+    "build_suggest_context",
+    "build_summarize_context",
+    "resolve_effective_current_line",
+]
