+++ plugin/plugins/galgame_plugin/models/__init__.py
@@ -1,0 +1,259 @@
+from __future__ import annotations
+
+from .config import (
+    GalgameBridgeConfig,
+    GalgameConfig,
+    GalgameHistoryConfig,
+    GalgameLLMConfig,
+    GalgameMemoryReaderConfig,
+    GalgameOcrReaderConfig,
+    GalgameRapidOcrConfig,
+    GalgameReaderConfig,
+    _ConfigFieldProxy,
+)
+from .constants import (
+    ADVANCE_SPEED_FAST,
+    ADVANCE_SPEED_MEDIUM,
+    ADVANCE_SPEED_SLOW,
+    ADVANCE_SPEEDS,
+    AGENT_STATUS_ACTIVE,
+    AGENT_STATUS_ERROR,
+    AGENT_STATUS_STANDBY,
+    AGENT_STATUSES,
+    CONNECTION_STATES,
+    DATA_SOURCE_BRIDGE_SDK,
+    DATA_SOURCE_MEMORY_READER,
+    DATA_SOURCE_NONE,
+    DATA_SOURCE_OCR_READER,
+    DATA_SOURCES,
+    DEFAULT_OCR_CAPTURE_BOTTOM_INSET_RATIO,
+    DEFAULT_OCR_CAPTURE_LEFT_INSET_RATIO,
+    DEFAULT_OCR_CAPTURE_RIGHT_INSET_RATIO,
+    DEFAULT_OCR_CAPTURE_TOP_RATIO,
+    MENU_PREFIX_RE,
+    MODE_CHOICE_ADVISOR,
+    MODE_COMPANION,
+    MODE_SILENT,
+    MODES,
+    OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUCKET_ASPECT_NEAREST,
+    OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUCKET_EXACT,
+    OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUILTIN_PRESET,
+    OCR_CAPTURE_PROFILE_MATCH_SOURCE_CONFIG_DEFAULT,
+    OCR_CAPTURE_PROFILE_MATCH_SOURCE_PROCESS_FALLBACK,
+    OCR_CAPTURE_PROFILE_RATIO_KEYS,
+    OCR_CAPTURE_PROFILE_SAVE_SCOPE_PROCESS_FALLBACK,
+    OCR_CAPTURE_PROFILE_SAVE_SCOPE_WINDOW_BUCKET,
+    OCR_CAPTURE_PROFILE_SAVE_SCOPES,
+    OCR_CAPTURE_PROFILE_STAGE_CONFIG,
+    OCR_CAPTURE_PROFILE_STAGE_DEFAULT,
+    OCR_CAPTURE_PROFILE_STAGE_DIALOGUE,
+    OCR_CAPTURE_PROFILE_STAGE_GALLERY,
+    OCR_CAPTURE_PROFILE_STAGE_GAME_OVER,
+    OCR_CAPTURE_PROFILE_STAGE_MENU,
+    OCR_CAPTURE_PROFILE_STAGE_MINIGAME,
+    OCR_CAPTURE_PROFILE_STAGE_SAVE_LOAD,
+    OCR_CAPTURE_PROFILE_STAGE_TITLE,
+    OCR_CAPTURE_PROFILE_STAGE_TRANSITION,
+    OCR_CAPTURE_PROFILE_STAGES,
+    OCR_CAPTURE_PROFILE_WINDOW_BUCKETS_KEY,
+    OCR_TRIGGER_MODE_AFTER_ADVANCE,
+    OCR_TRIGGER_MODE_INTERVAL,
+    OCR_TRIGGER_MODES,
+    READER_MODE_AUTO,
+    READER_MODE_MEMORY,
+    READER_MODE_OCR,
+    READER_MODES,
+    STATE_ACTIVE,
+    STATE_DISCONNECTED,
+    STATE_ERROR,
+    STATE_IDLE,
+    STATE_STALE,
+    SharedStatePayload,
+)
+from .sanitizers import (
+    DEFAULT_SAVE_CONTEXT,
+    _bool,
+    _float,
+    _int,
+    _sanitize_choice_bounds,
+    _string,
+    build_ocr_capture_profile_bucket_key,
+    compute_ocr_window_aspect_ratio,
+    json_copy,
+    make_error,
+    normalize_rapidocr_ocr_version,
+    parse_ocr_capture_profile_bucket_key,
+    sanitize_choice,
+    sanitize_event,
+    sanitize_metadata,
+    sanitize_save_context,
+    sanitize_screen_ui_element,
+    sanitize_screen_ui_elements,
+    sanitize_session_snapshot,
+    sanitize_snapshot_state,
+)
+from .session import SessionCandidate
+from .store_keys import (
+    STORE_ADVANCE_SPEED,
+    STORE_BOUND_GAME_ID,
+    STORE_CHARACTER_FIXED_NAME,
+    STORE_CHARACTER_MODE,
+    STORE_CHARACTER_PROFILE_VERSION,
+    STORE_CHARACTER_PROFILES,
+    STORE_CHARACTER_RUNTIME_STATE,
+    STORE_CONTEXT_SNAPSHOT,
+    STORE_CROSS_SCENE_MEMORY,
+    STORE_DEDUPE_WINDOW,
+    STORE_EVENTS_BYTE_OFFSET,
+    STORE_EVENTS_FILE_SIZE,
+    STORE_KEYS,
+    STORE_LAST_ERROR,
+    STORE_LAST_SEQ,
+    STORE_LLM_VISION_ENABLED,
+    STORE_LLM_VISION_MAX_IMAGE_PX,
+    STORE_MEMORY_READER_TARGET,
+    STORE_MODE,
+    STORE_OCR_BACKEND_SELECTION,
+    STORE_OCR_CAPTURE_BACKEND,
+    STORE_OCR_CAPTURE_PROFILES,
+    STORE_OCR_FAST_LOOP_ENABLED,
+    STORE_OCR_POLL_INTERVAL_SECONDS,
+    STORE_OCR_SCREEN_TEMPLATES,
+    STORE_OCR_TRIGGER_MODE,
+    STORE_OCR_WINDOW_TARGET,
+    STORE_PUSH_NOTIFICATIONS,
+    STORE_RAPIDOCR_AUTO_DETECT_LANG,
+    STORE_RAPIDOCR_AUTO_DETECT_LAST_LANG,
+    STORE_RAPIDOCR_LANG_TYPE,
+    STORE_RAPIDOCR_OCR_VERSION,
+    STORE_READER_MODE,
+    STORE_SESSION_ID,
+    STORE_TUTORIAL_PROGRESS,
+    _RAPIDOCR_OCR_VERSIONS,
+)
+
+__all__ = [
+    "ADVANCE_SPEED_FAST",
+    "ADVANCE_SPEED_MEDIUM",
+    "ADVANCE_SPEED_SLOW",
+    "ADVANCE_SPEEDS",
+    "AGENT_STATUS_ACTIVE",
+    "AGENT_STATUS_ERROR",
+    "AGENT_STATUS_STANDBY",
+    "AGENT_STATUSES",
+    "CONNECTION_STATES",
+    "DATA_SOURCE_BRIDGE_SDK",
+    "DATA_SOURCE_MEMORY_READER",
+    "DATA_SOURCE_NONE",
+    "DATA_SOURCE_OCR_READER",
+    "DATA_SOURCES",
+    "DEFAULT_OCR_CAPTURE_BOTTOM_INSET_RATIO",
+    "DEFAULT_OCR_CAPTURE_LEFT_INSET_RATIO",
+    "DEFAULT_OCR_CAPTURE_RIGHT_INSET_RATIO",
+    "DEFAULT_OCR_CAPTURE_TOP_RATIO",
+    "DEFAULT_SAVE_CONTEXT",
+    "GalgameBridgeConfig",
+    "GalgameConfig",
+    "GalgameHistoryConfig",
+    "GalgameLLMConfig",
+    "GalgameMemoryReaderConfig",
+    "GalgameOcrReaderConfig",
+    "GalgameRapidOcrConfig",
+    "GalgameReaderConfig",
+    "MENU_PREFIX_RE",
+    "MODE_CHOICE_ADVISOR",
+    "MODE_COMPANION",
+    "MODE_SILENT",
+    "MODES",
+    "OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUCKET_ASPECT_NEAREST",
+    "OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUCKET_EXACT",
+    "OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUILTIN_PRESET",
+    "OCR_CAPTURE_PROFILE_MATCH_SOURCE_CONFIG_DEFAULT",
+    "OCR_CAPTURE_PROFILE_MATCH_SOURCE_PROCESS_FALLBACK",
+    "OCR_CAPTURE_PROFILE_RATIO_KEYS",
+    "OCR_CAPTURE_PROFILE_SAVE_SCOPE_PROCESS_FALLBACK",
+    "OCR_CAPTURE_PROFILE_SAVE_SCOPE_WINDOW_BUCKET",
+    "OCR_CAPTURE_PROFILE_SAVE_SCOPES",
+    "OCR_CAPTURE_PROFILE_STAGE_CONFIG",
+    "OCR_CAPTURE_PROFILE_STAGE_DEFAULT",
+    "OCR_CAPTURE_PROFILE_STAGE_DIALOGUE",
+    "OCR_CAPTURE_PROFILE_STAGE_GALLERY",
+    "OCR_CAPTURE_PROFILE_STAGE_GAME_OVER",
+    "OCR_CAPTURE_PROFILE_STAGE_MENU",
+    "OCR_CAPTURE_PROFILE_STAGE_MINIGAME",
+    "OCR_CAPTURE_PROFILE_STAGE_SAVE_LOAD",
+    "OCR_CAPTURE_PROFILE_STAGE_TITLE",
+    "OCR_CAPTURE_PROFILE_STAGE_TRANSITION",
+    "OCR_CAPTURE_PROFILE_STAGES",
+    "OCR_CAPTURE_PROFILE_WINDOW_BUCKETS_KEY",
+    "OCR_TRIGGER_MODE_AFTER_ADVANCE",
+    "OCR_TRIGGER_MODE_INTERVAL",
+    "OCR_TRIGGER_MODES",
+    "READER_MODE_AUTO",
+    "READER_MODE_MEMORY",
+    "READER_MODE_OCR",
+    "READER_MODES",
+    "STATE_ACTIVE",
+    "STATE_DISCONNECTED",
+    "STATE_ERROR",
+    "STATE_IDLE",
+    "STATE_STALE",
+    "STORE_ADVANCE_SPEED",
+    "STORE_BOUND_GAME_ID",
+    "STORE_CHARACTER_FIXED_NAME",
+    "STORE_CHARACTER_MODE",
+    "STORE_CHARACTER_PROFILE_VERSION",
+    "STORE_CHARACTER_PROFILES",
+    "STORE_CHARACTER_RUNTIME_STATE",
+    "STORE_CONTEXT_SNAPSHOT",
+    "STORE_CROSS_SCENE_MEMORY",
+    "STORE_DEDUPE_WINDOW",
+    "STORE_EVENTS_BYTE_OFFSET",
+    "STORE_EVENTS_FILE_SIZE",
+    "STORE_KEYS",
+    "STORE_LAST_ERROR",
+    "STORE_LAST_SEQ",
+    "STORE_LLM_VISION_ENABLED",
+    "STORE_LLM_VISION_MAX_IMAGE_PX",
+    "STORE_MEMORY_READER_TARGET",
+    "STORE_MODE",
+    "STORE_OCR_BACKEND_SELECTION",
+    "STORE_OCR_CAPTURE_BACKEND",
+    "STORE_OCR_CAPTURE_PROFILES",
+    "STORE_OCR_FAST_LOOP_ENABLED",
+    "STORE_OCR_POLL_INTERVAL_SECONDS",
+    "STORE_OCR_SCREEN_TEMPLATES",
+    "STORE_OCR_TRIGGER_MODE",
+    "STORE_OCR_WINDOW_TARGET",
+    "STORE_PUSH_NOTIFICATIONS",
+    "STORE_RAPIDOCR_AUTO_DETECT_LANG",
+    "STORE_RAPIDOCR_AUTO_DETECT_LAST_LANG",
+    "STORE_RAPIDOCR_LANG_TYPE",
+    "STORE_RAPIDOCR_OCR_VERSION",
+    "STORE_READER_MODE",
+    "STORE_SESSION_ID",
+    "STORE_TUTORIAL_PROGRESS",
+    "SessionCandidate",
+    "SharedStatePayload",
+    "_ConfigFieldProxy",
+    "_bool",
+    "_float",
+    "_int",
+    "_sanitize_choice_bounds",
+    "_string",
+    "build_ocr_capture_profile_bucket_key",
+    "compute_ocr_window_aspect_ratio",
+    "json_copy",
+    "make_error",
+    "normalize_rapidocr_ocr_version",
+    "parse_ocr_capture_profile_bucket_key",
+    "sanitize_choice",
+    "sanitize_event",
+    "sanitize_metadata",
+    "sanitize_save_context",
+    "sanitize_screen_ui_element",
+    "sanitize_screen_ui_elements",
+    "sanitize_session_snapshot",
+    "sanitize_snapshot_state",
+    "_RAPIDOCR_OCR_VERSIONS",
+]
