+++ plugin/plugins/galgame_plugin/models/constants.py
@@ -1,0 +1,103 @@
+from __future__ import annotations
+
+import re
+from typing import Any, TypeAlias
+
+MODE_SILENT = "silent"
+MODE_COMPANION = "companion"
+MODE_CHOICE_ADVISOR = "choice_advisor"
+MODES = frozenset({MODE_SILENT, MODE_COMPANION, MODE_CHOICE_ADVISOR})
+
+ADVANCE_SPEED_SLOW = "slow"
+ADVANCE_SPEED_MEDIUM = "medium"
+ADVANCE_SPEED_FAST = "fast"
+ADVANCE_SPEEDS = frozenset({ADVANCE_SPEED_SLOW, ADVANCE_SPEED_MEDIUM, ADVANCE_SPEED_FAST})
+
+DATA_SOURCE_NONE = "none"
+DATA_SOURCE_BRIDGE_SDK = "bridge_sdk"
+DATA_SOURCE_MEMORY_READER = "memory_reader"
+DATA_SOURCE_OCR_READER = "ocr_reader"
+DATA_SOURCES = frozenset(
+    {
+        DATA_SOURCE_NONE,
+        DATA_SOURCE_BRIDGE_SDK,
+        DATA_SOURCE_MEMORY_READER,
+        DATA_SOURCE_OCR_READER,
+    }
+)
+SharedStatePayload: TypeAlias = dict[str, Any]
+MENU_PREFIX_RE = re.compile(r"^\s*(?:[-*•]\s+|\d+[\.\)\]:：]\s+)(.+\S)\s*$")
+
+READER_MODE_AUTO = "auto"
+READER_MODE_MEMORY = DATA_SOURCE_MEMORY_READER
+READER_MODE_OCR = DATA_SOURCE_OCR_READER
+READER_MODES = frozenset({READER_MODE_AUTO, READER_MODE_MEMORY, READER_MODE_OCR})
+
+OCR_TRIGGER_MODE_INTERVAL = "interval"
+OCR_TRIGGER_MODE_AFTER_ADVANCE = "after_advance"
+OCR_TRIGGER_MODES = frozenset({OCR_TRIGGER_MODE_INTERVAL, OCR_TRIGGER_MODE_AFTER_ADVANCE})
+
+DEFAULT_OCR_CAPTURE_LEFT_INSET_RATIO = 0.05
+DEFAULT_OCR_CAPTURE_RIGHT_INSET_RATIO = 0.05
+DEFAULT_OCR_CAPTURE_TOP_RATIO = 0.62
+DEFAULT_OCR_CAPTURE_BOTTOM_INSET_RATIO = 0.08
+OCR_CAPTURE_PROFILE_RATIO_KEYS = (
+    "left_inset_ratio",
+    "right_inset_ratio",
+    "top_ratio",
+    "bottom_inset_ratio",
+)
+OCR_CAPTURE_PROFILE_STAGE_DEFAULT = "default"
+OCR_CAPTURE_PROFILE_STAGE_DIALOGUE = "dialogue_stage"
+OCR_CAPTURE_PROFILE_STAGE_MENU = "menu_stage"
+OCR_CAPTURE_PROFILE_STAGE_TITLE = "title_stage"
+OCR_CAPTURE_PROFILE_STAGE_SAVE_LOAD = "save_load_stage"
+OCR_CAPTURE_PROFILE_STAGE_CONFIG = "config_stage"
+OCR_CAPTURE_PROFILE_STAGE_TRANSITION = "transition_stage"
+OCR_CAPTURE_PROFILE_STAGE_GALLERY = "gallery_stage"
+OCR_CAPTURE_PROFILE_STAGE_MINIGAME = "minigame_stage"
+OCR_CAPTURE_PROFILE_STAGE_GAME_OVER = "game_over_stage"
+OCR_CAPTURE_PROFILE_STAGES = frozenset(
+    {
+        OCR_CAPTURE_PROFILE_STAGE_DEFAULT,
+        OCR_CAPTURE_PROFILE_STAGE_DIALOGUE,
+        OCR_CAPTURE_PROFILE_STAGE_MENU,
+        OCR_CAPTURE_PROFILE_STAGE_TITLE,
+        OCR_CAPTURE_PROFILE_STAGE_SAVE_LOAD,
+        OCR_CAPTURE_PROFILE_STAGE_CONFIG,
+        OCR_CAPTURE_PROFILE_STAGE_TRANSITION,
+        OCR_CAPTURE_PROFILE_STAGE_GALLERY,
+        OCR_CAPTURE_PROFILE_STAGE_MINIGAME,
+        OCR_CAPTURE_PROFILE_STAGE_GAME_OVER,
+    }
+)
+OCR_CAPTURE_PROFILE_WINDOW_BUCKETS_KEY = "__window_buckets__"
+OCR_CAPTURE_PROFILE_SAVE_SCOPE_WINDOW_BUCKET = "window_bucket"
+OCR_CAPTURE_PROFILE_SAVE_SCOPE_PROCESS_FALLBACK = "process_fallback"
+OCR_CAPTURE_PROFILE_SAVE_SCOPES = frozenset(
+    {
+        OCR_CAPTURE_PROFILE_SAVE_SCOPE_WINDOW_BUCKET,
+        OCR_CAPTURE_PROFILE_SAVE_SCOPE_PROCESS_FALLBACK,
+    }
+)
+OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUCKET_EXACT = "bucket_exact"
+OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUCKET_ASPECT_NEAREST = "bucket_aspect_nearest"
+OCR_CAPTURE_PROFILE_MATCH_SOURCE_PROCESS_FALLBACK = "process_fallback"
+OCR_CAPTURE_PROFILE_MATCH_SOURCE_BUILTIN_PRESET = "builtin_preset"
+OCR_CAPTURE_PROFILE_MATCH_SOURCE_CONFIG_DEFAULT = "config_default"
+
+AGENT_STATUS_ACTIVE = "active"
+AGENT_STATUS_STANDBY = "standby"
+AGENT_STATUS_ERROR = "error"
+AGENT_STATUSES = frozenset(
+    {AGENT_STATUS_ACTIVE, AGENT_STATUS_STANDBY, AGENT_STATUS_ERROR}
+)
+
+STATE_DISCONNECTED = "disconnected"
+STATE_IDLE = "idle"
+STATE_ACTIVE = "active"
+STATE_STALE = "stale"
+STATE_ERROR = "error"
+CONNECTION_STATES = frozenset(
+    {STATE_DISCONNECTED, STATE_IDLE, STATE_ACTIVE, STATE_STALE, STATE_ERROR}
+)
