+++ plugin/plugins/galgame_plugin/models/store_keys.py
@@ -1,0 +1,73 @@
+from __future__ import annotations
+
+STORE_BOUND_GAME_ID = "bound_game_id"
+STORE_MODE = "mode"
+STORE_PUSH_NOTIFICATIONS = "push_notifications"
+STORE_ADVANCE_SPEED = "advance_speed"
+STORE_SESSION_ID = "session_id"
+STORE_EVENTS_BYTE_OFFSET = "events_byte_offset"
+STORE_EVENTS_FILE_SIZE = "events_file_size"
+STORE_LAST_SEQ = "last_seq"
+STORE_DEDUPE_WINDOW = "dedupe_window"
+STORE_LAST_ERROR = "last_error"
+STORE_OCR_CAPTURE_PROFILES = "ocr_capture_profiles"
+STORE_OCR_WINDOW_TARGET = "ocr_window_target"
+STORE_MEMORY_READER_TARGET = "memory_reader_target"
+STORE_OCR_BACKEND_SELECTION = "ocr_backend_selection"
+STORE_OCR_CAPTURE_BACKEND = "ocr_capture_backend"
+STORE_READER_MODE = "reader_mode"
+STORE_OCR_POLL_INTERVAL_SECONDS = "ocr_poll_interval_seconds"
+STORE_OCR_TRIGGER_MODE = "ocr_trigger_mode"
+STORE_LLM_VISION_ENABLED = "llm_vision_enabled"
+STORE_LLM_VISION_MAX_IMAGE_PX = "llm_vision_max_image_px"
+STORE_OCR_SCREEN_TEMPLATES = "ocr_screen_templates"
+STORE_OCR_FAST_LOOP_ENABLED = "ocr_fast_loop_enabled"
+STORE_RAPIDOCR_LANG_TYPE = "rapidocr.lang_type"
+STORE_RAPIDOCR_AUTO_DETECT_LANG = "rapidocr.auto_detect_lang"
+STORE_RAPIDOCR_AUTO_DETECT_LAST_LANG = "rapidocr.auto_detect_last_lang"
+STORE_RAPIDOCR_OCR_VERSION = "rapidocr.ocr_version"
+_RAPIDOCR_OCR_VERSIONS: frozenset[str] = frozenset({"PP-OCRv4", "PP-OCRv5"})
+STORE_TUTORIAL_PROGRESS = "tutorial_progress"
+STORE_CONTEXT_SNAPSHOT = "context_snapshot"
+STORE_CHARACTER_PROFILES = "character_profiles"
+STORE_CHARACTER_PROFILE_VERSION = "character_profile_version"
+STORE_CHARACTER_MODE = "character_mode"
+STORE_CHARACTER_FIXED_NAME = "character_fixed_name"
+STORE_CROSS_SCENE_MEMORY = "cross_scene_memory"
+STORE_CHARACTER_RUNTIME_STATE = "character_runtime_state"
+STORE_KEYS = (
+    STORE_BOUND_GAME_ID,
+    STORE_MODE,
+    STORE_PUSH_NOTIFICATIONS,
+    STORE_ADVANCE_SPEED,
+    STORE_SESSION_ID,
+    STORE_EVENTS_BYTE_OFFSET,
+    STORE_EVENTS_FILE_SIZE,
+    STORE_LAST_SEQ,
+    STORE_DEDUPE_WINDOW,
+    STORE_LAST_ERROR,
+    STORE_OCR_CAPTURE_PROFILES,
+    STORE_OCR_WINDOW_TARGET,
+    STORE_MEMORY_READER_TARGET,
+    STORE_OCR_BACKEND_SELECTION,
+    STORE_OCR_CAPTURE_BACKEND,
+    STORE_READER_MODE,
+    STORE_OCR_POLL_INTERVAL_SECONDS,
+    STORE_OCR_TRIGGER_MODE,
+    STORE_LLM_VISION_ENABLED,
+    STORE_LLM_VISION_MAX_IMAGE_PX,
+    STORE_OCR_SCREEN_TEMPLATES,
+    STORE_OCR_FAST_LOOP_ENABLED,
+    STORE_RAPIDOCR_LANG_TYPE,
+    STORE_RAPIDOCR_AUTO_DETECT_LANG,
+    STORE_RAPIDOCR_AUTO_DETECT_LAST_LANG,
+    STORE_RAPIDOCR_OCR_VERSION,
+    STORE_TUTORIAL_PROGRESS,
+    STORE_CONTEXT_SNAPSHOT,
+    STORE_CHARACTER_PROFILES,
+    STORE_CHARACTER_PROFILE_VERSION,
+    STORE_CHARACTER_MODE,
+    STORE_CHARACTER_FIXED_NAME,
+    STORE_CROSS_SCENE_MEMORY,
+    STORE_CHARACTER_RUNTIME_STATE,
+)
