+++ plugin/plugins/galgame_plugin/store/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from .core import GalgameStore
+
+__all__ = ["GalgameStore"]
